==> 01_data_scrape.Rmd <==
---
title: "SFPA Ratings - Data Scraping"
author: "Skip Perry"
date: "January 2020"
output: github_document
---
  
```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(rvest)
library(lubridate)
```

```{r}
# Produces a list of URLs of score sheets for a given completed week
# Starts with a URL of form week_url <- "https://www.sfpapool.org/stats/week/86/"
# Grabs the first scoresheet in each row, so need to wait until results finalized

get_week_match_urls <- function(week_url) {
  week_url %>% 
    read_html() %>% 
    html_nodes("tr") %>% 
    html_nodes("td") %>% 
    html_node("a") %>% 
    html_attr("href") %>% 
    enframe(name = NULL) %>% 
    filter(str_detect(value, "score_sheet")) %>% 
    transmute(score_sheets = str_c("https://www.sfpapool.org", value)) %>% 
    pull(score_sheets)
}

#get_week_match_urls("https://www.sfpapool.org/stats/week/58/")

# Parses a given score sheet and turns it into a data frame
# Score sheet has form https://www.sfpapool.org/stats/score_sheet/1223/

parse_score_sheet <- function(url) {
  #url <- 'https://www.sfpapool.org/stats/score_sheet/1799/'

  webpage <- read_html(url)
  
  match_date <-
    webpage %>% 
    html_nodes("h5") %>% 
    html_text() %>% 
    enframe(name = NULL) %>% 
    slice(2) %>% 
    pull() %>% 
    str_replace("Sept.", "Sep.") %>% 
    str_extract("\\w+.\\s\\d+,\\s\\d+") %>% 
    mdy()

  season <- 
    str_c(
      ifelse(match_date %>% month < 7, "Spring", "Fall"),
      match_date %>% year, 
      sep = " "
    )
  
  week_number <- 
    webpage %>% 
    html_nodes("h5") %>% 
    html_text %>% 
    enframe(name = NULL) %>% 
    filter(str_detect(value, pattern = "Week")) %>% 
    as.character() %>% 
    str_extract(pattern = "\\d+") %>% 
    as.numeric()
  
  # Playoff games are labeled 50
  week_number <- ifelse(week_number == 0, 50, week_number)
  
  away_team <- 
    webpage %>% 
    html_nodes("h4") %>% 
    html_text() %>% 
    enframe(name = NULL) %>% 
    filter(str_detect(value, "Summaries")) %>% 
    slice(1) %>% 
    as.character() %>% 
    str_remove("\\sSummaries")
  
  home_team <- 
    webpage %>% 
    html_nodes("h4") %>% 
    html_text() %>% 
    enframe(name = NULL) %>% 
    filter(str_detect(value, "Summaries")) %>% 
    slice(2) %>% 
    as.character() %>% 
    str_remove("\\sSummaries")
  
  game_results <- 
    webpage %>% 
    html_nodes("tr") %>% 
    html_nodes("td") %>% 
    html_nodes("div") %>% 
    html_nodes("input") %>% 
    html_attr("checked") %>% 
    enframe(name = NULL) %>% 
    mutate(
      team = rep(c("home", "away"), nrow(.) / 2),
      game_num = rep(1:(nrow(.) / 2), each = 2)
    ) %>% 
    spread(key = team, value = value) %>% 
    transmute(
      game_winner = case_when(
        !is.na(home) ~ "home",
        !is.na(away) ~ "away"
      )
    )
  
  player_matchups <-
    webpage %>% 
    html_nodes("div") %>% 
    html_nodes("div") %>% 
    html_nodes("div") %>% 
    html_text() %>% 
    enframe(name = NULL) %>% 
    filter(str_detect(value, "\\w+\\s\\w+|\\w+|--")) %>% 
    filter(!str_detect(value, "^\n")) %>% 
    filter(!str_detect(value, "^\\d+")) %>% 
    filter(value != "A") %>% 
    filter(value != "B") %>% 
    filter(value != "C") %>% 
    filter(value != "D") %>% 
    filter(value != "Home TB") %>% 
    filter(value != "Away TB") %>% 
    mutate(
      value = if_else(value == "--", "Forfeited Game", value),
      team = rep(c("home", "away"), nrow(.) / 2),
      game_num = rep(1:(nrow(.) / 2), each = 2)
    ) %>% 
    spread(key = "team", value = "value") %>% 
    bind_cols(game_results) %>% 
    select(game_num, home, away, game_winner)
  
  forfeited_games <- 
    webpage %>% 
    html_nodes("tr") %>% 
    html_nodes("td") %>% 
    toString() %>% 
    str_extract_all(pattern = 'form-\\d+-forfeit\\"\\schecked') %>% 
    unlist() %>% 
    enframe(name = NULL) %>% 
    transmute(
      game_num = as.numeric(str_extract(value, pattern = "\\d+")) + 1,
      forfeit = "forfeit"
    )
  
  #time_completed <- 
  #  webpage %>% 
  #  html_nodes("tr") %>% 
  #  html_nodes("td") %>% 
  #  toString() %>% 
  #  str_extract_all(pattern = '\\d+\\:\\d+') %>% 
  #  unlist() %>% 
  #  enframe(name = NULL) %>% 
  #  transmute(
  #    game_num = row_number(),
  #    time_completed = value
  #  )

  match_summary <- 
    player_matchups %>% 
    mutate(
      season = season,
      match_date = match_date,
      week_number = week_number,
      home_team = home_team,
      away_team = away_team
    ) %>% 
    select(season, match_date, week_number, home_team, away_team, game_num:game_winner) %>% 
    left_join(forfeited_games, by = "game_num")
    #left_join(time_completed, by = "game_num")
  
  match_summary
}

#parse_score_sheet("https://www.sfpapool.org/stats/score_sheet/1799/")

# Function to turn a list of week URLs into a data frame of match results for a whole season
# Cycles through each week URL, grabs the matches from that week, adds to data frame
week_urls_to_parsed_matches <- function(week_urls) {
  parsed_matches <- tibble() 

  for (i in 1:length(week_urls)) {
    print(str_c("Starting week ", i))
    match_urls <- get_week_match_urls(week_urls[i])
    week_match_data <- map_dfr(match_urls, parse_score_sheet)
    parsed_matches <- bind_rows(parsed_matches, week_match_data)
    print(str_c("Week ", i, " complete"))
  }
  
  parsed_matches
}

#week_urls_to_parsed_matches(str_c("https://www.sfpapool.org/stats/week/", c(57:58)))
```

```{r}
# Summarize previous file
prev_matches <- readRDS("match_data/all_matches_2020spring.Rdata")

prev_matches %>% 
  group_by(match_date) %>% 
  count()
```

```{r}
# Get updated Spring 2020 results and check
weeks_to_parse <- str_c("https://www.sfpapool.org/stats/week/", 107) # Most recent week
new_matches <- week_urls_to_parsed_matches(weeks_to_parse)

bind_rows(prev_matches, new_matches) %>% 
  group_by(match_date) %>% 
  count()
```

```{r}
# Add to previous results
all_matches_2020spring <- bind_rows(prev_matches, new_matches)

saveRDS(all_matches_2020spring, "match_data/all_matches_2020spring.Rdata")
```

```{r, eval=FALSE, echo=FALSE}
#Scrape an outside file from a tournament:
prev_matches <- readRDS("match_data/all_matches_2019springplayoffs.Rdata")

prev_matches %>% 
  group_by(match_date) %>% 
  count()

new_matches <- 
  read_csv("tournaments/spring2019_slate8ball_week1.csv") %>% 
  mutate(home_team = NA_character_, away_team = NA_character_, forfeit = NA_character_)

bind_rows(prev_matches, new_matches) %>% 
  group_by(match_date) %>% 
  count()

# Add to previous results
all_matches_2019springplayoffs <- bind_rows(prev_matches, new_matches)
  
saveRDS(all_matches_2019springplayoffs, "match_data/all_matches_2019springplayoffs.Rdata")
```

```{r, eval=FALSE, echo=FALSE}
# Spring 2018 data generation:

# Save down 2018 spring results
week_urls_2018spring <- str_c("https://www.sfpapool.org/stats/week/", c(9:10)) # Spring 2018
all_matches_2018spring <- week_urls_to_all_matches(week_urls_2018spring)

all_matches_2018spring %>% 
  group_by(game_num) %>% 
  count()

all_matches_2018spring %>% 
  group_by(match_date) %>% 
  count()

saveRDS(all_matches_2018spring, "match_data/all_matches_2018spring.Rdata")

week_urls_2018springplayoffs <- str_c("https://www.sfpapool.org/stats/week/", c(23:32)) # Spring 2018 playoffs
all_matches_2018springplayoffs <- week_urls_to_all_matches(week_urls_2018springplayoffs)

all_matches_2018springplayoffs %>% 
  group_by(game_num) %>% 
  count()

all_matches_2018springplayoffs %>% 
  group_by(match_date) %>% 
  count()

saveRDS(all_matches_2018springplayoffs, "match_data/all_matches_2018springplayoffs.Rdata")
```

```{r, eval=FALSE, echo=FALSE}
# Fall 2018 data generation:

# Save down 2018 fall results
week_urls_2018fall <- str_c("https://www.sfpapool.org/stats/week/", c(33:39, 41:46)) # Fall 2018
all_matches_2018fall <- week_urls_to_all_matches(week_urls_2018fall)

all_matches_2018fall %>% 
  group_by(game_num) %>% 
  count()

all_matches_2018fall %>% 
  group_by(match_date) %>% 
  count()

saveRDS(all_matches_2018fall, "match_data/all_matches_2018fall.Rdata")

week_urls_2018fallplayoffs <- str_c("https://www.sfpapool.org/stats/week/", c(48:56)) # Fall 2018 playoffs
all_matches_2018fallplayoffs <- week_urls_to_all_matches(week_urls_2018fallplayoffs)

all_matches_2018fallplayoffs %>% 
  group_by(game_num) %>% 
  count()

all_matches_2018fallplayoffs %>% 
  group_by(match_date) %>% 
  count()

saveRDS(all_matches_2018fallplayoffs, "match_data/all_matches_2018fallplayoffs.Rdata")
```

```{r, eval=FALSE, echo=FALSE}
webpage %>% 
  html_nodes(xpath = "/html/body/div[2]/div/div/table/thead/tr[2]/th[2]/h4/a") %>% 
  .[[1]]
```
