==> 03_update_ratings.Rmd <==
---
title: "SFPA Ratings v.1"
author: "Skip Perry"
date: "January 2020"
output: github_document
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(glue)
```

```{r}
latest_results_date <- 
  list.files("match_data", pattern = "results_no_forfeits") %>% 
  str_extract("\\d+-\\d+-\\d+") %>% 
  max()

results_no_forfeits <- 
  str_c("match_data/results_no_forfeits_", latest_results_date, ".Rdata") %>% 
  read_rds()

# Computes udpated ratings for all players until the absolute difference 
# in the vector of ratings between iterations reaches the desired threshold
# Usually takes about 20-30 iterations to converge
results_to_ratings <- function(results_df, a, stop_value = 100) {
  start_time <- Sys.time()

  # Set up initial ratings at 500
  fargo_ratings <- 
    bind_rows(
      results_df %>% transmute(player = home), 
      results_df %>% transmute(player = away)
    ) %>% 
    distinct() %>% 
    arrange(player) %>% 
    mutate(rating = 500)
  
  # Creates a data frame with columns for season-player-opponent-result-decay 
  # for a single player. This definition is inside the function because it uses 
  # results_df which is passed into results_to_ratings.
  collect_game_results <- function(player_of_interest) {
    results_df %>% 
      filter(away == player_of_interest | home == player_of_interest) %>% 
      transmute(
        season,
        match_date,
        player = player_of_interest,
        opponent = if_else(home == player_of_interest, away, home),
        game_result = case_when(
          home == player_of_interest & game_winner == "home" ~ "W",
          away == player_of_interest & game_winner == "away" ~ "W",
          TRUE ~ "L"
        )
      )
  }

  # This un-tidy data frame has 2 rows for every game played. For each game, 
  # one row has player 1 as the target and another row has player 2 as the 
  # target. Results can then be grouped by player and produce a fast,
  # vectorized rating update after joining to the latest rating predictions. 
  # Adds decay step factor of 5% for every six months from the latest game. 
  collected_game_results <- 
    map_dfr(fargo_ratings %>% pull(player), collect_game_results) %>% 
    mutate(
      max_date = max(match_date),
      date_diff = interval(max_date, match_date) %/% months(6),
      decay = 1 + date_diff * 0.05
    ) %>% 
    select(-max_date, -date_diff)
  
  # Set up convergence and MAP parameters 
  abs_diff <- 1e5
  n_iter <- 0
  b <- (a - 1) / 500
  
  # Update ratings using iterative MAP algorithm from 05_info_faq.md
  while (abs_diff > stop_value) {
    n_iter <- n_iter + 1
    
    # Existing vector of player ratings for convergence test
    old_ratings <- fargo_ratings %>% pull(rating)

    # Update rating estimates
    fargo_ratings <- 
      # Start with collected results, arranged by player 
      collected_game_results %>%
      # Add current player ratings
      inner_join(
        fargo_ratings %>% transmute(player, player_rating = rating), 
        by = "player"
      ) %>% 
      # Add current opponent ratings
      inner_join(
        fargo_ratings %>% transmute(opponent = player, opponent_rating = rating), 
        by = "opponent"
      ) %>% 
      # Implement MAP algorithm 
      mutate(
        A = 1 / (player_rating + opponent_rating),
        W = if_else(game_result == "W", 1, 0)
      ) %>% 
      group_by(player) %>% 
      summarize(rating = ((a - 1) + sum(W * decay)) / (b + sum(A * decay)))

    # New vector of player ratings for convergence test
    new_ratings <- fargo_ratings %>% pull(rating) 
    
    print(str_c("Sum of absolute difference: ", sum(abs(old_ratings - new_ratings))))
    
    # Break out of loop if abs_diff < stop_value
    abs_diff <- sum(abs(old_ratings - new_ratings))
  }
  end_time <- Sys.time()
  time_taken <- round(end_time - start_time, 2)

  print(glue("Number of iterations: {n_iter}"))
  print(glue("Time taken: {time_taken} seconds"))
  
  fargo_ratings
}
```

```{r}
fargo_df <- 
  results_to_ratings(results_df = results_no_forfeits, a = 3, stop_value = 50) %>% 
  mutate(
    raw_rating = rating,
    rating = log(rating) * 144,
    rating = rating - mean(rating) + 500
  )

fargo_df %>% 
  arrange(desc(rating))
```

```{r}
saveRDS(fargo_df, str_c("fargo_ratings/fargo_", latest_results_date, ".Rdata"))
```

```{r}
# Create a single file with all known player ratings
fargo_files <- 
  list.files("fargo_ratings", pattern = "^fargo")

fargo_dates <- 
  fargo_files %>% 
  str_extract("\\d+-\\d+-\\d+") %>% 
  enframe(name = NULL) %>% 
  arrange(value) %>% 
  transmute(
    source = row_number(),
    date = ymd(value)
  )

all_fargo_ratings <- 
  str_c("fargo_ratings/", fargo_files) %>% 
  map_dfr(read_rds, .id = "source") %>% 
  mutate(source = as.numeric(source)) %>% 
  left_join(fargo_dates, by = "source") %>% 
  select(-source)

saveRDS(all_fargo_ratings, str_c("fargo_ratings/all_fargo_ratings.Rdata"))
```

```{r, echo=FALSE, eval=FALSE}
# Do this to update previous fargo ratings after adding new games to database
# from before current week, making changes to player names, etc.
date_list <- 
  results_no_forfeits %>% 
  select(match_date) %>% 
  distinct() %>% 
  arrange(match_date) %>% 
  pull()

num_dates <- 126
# Last time ran back to beginning of 2018

#length(date_list)
tail(date_list, num_dates)

#for (i in 71:102) {
for (i in 1:num_dates) {
  results_date <- tail(date_list, num_dates)[i]
  
  print(results_date)
  
  fargo_df <- 
    results_to_ratings(
      results_df = results_no_forfeits %>% filter(match_date <= results_date), 
      a = 3, 
      stop_value = 50
    ) %>% 
    mutate(
      raw_rating = rating,
      rating = log(rating) * 144,
      rating = rating - mean(rating) + 500
    )
  
  saveRDS(fargo_df, str_c("fargo_ratings/fargo_", results_date, ".Rdata"))
}
```

```{r, eval=FALSE, echo=FALSE}
# SCRATCH FROM CHANGING DECAY
results_df <- results_no_forfeits
player_of_interest <- "Patty West"

results_df %>% 
  filter(away == player_of_interest | home == player_of_interest) %>% 
  transmute(
    season,
    player = player_of_interest,
    opponent = if_else(home == player_of_interest, away, home),
    game_result = case_when(
      home == player_of_interest & game_winner == "home" ~ "W",
      away == player_of_interest & game_winner == "away" ~ "W",
      TRUE ~ "L"
    ),
    match_date
  )
    decay = case_when(
      season == "Spring 2012" ~ 0.3,
      season == "Fall 2012" ~ 0.35,
      season == "Spring 2013" ~ 0.4,
      season == "Fall 2013" ~ 0.45,
      season == "Spring 2014" ~ 0.5,
      season == "Fall 2014" ~ 0.55,
      season == "Spring 2015" ~ 0.6,
      season == "Fall 2015" ~ 0.65,
      season == "Spring 2016" ~ 0.7,
      season == "Fall 2016" ~ 0.75,
      season == "Spring 2017" ~ 0.8,
      season == "Fall 2017" ~ 0.85,
      season == "Spring 2018" ~ 0.9,
      season == "Fall 2018" ~ 0.95,
      season == "Spring 2019" ~ 1.0,
      season == "Fall 2019" ~ 1.0,
      season == "Spring 2020" ~ 1.0
    )
  )
```

```{r}
```
