==> 06_ratings_over_time.Rmd <==
---
title: "SFPA Player Ratings, v.1"
author: "Skip Perry"
date: "January 2020"
output: github_document
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)

latest_results_date <- 
  list.files("match_data", pattern = "results_no_forfeits") %>% 
  str_extract("\\d+-\\d+-\\d+") %>% 
  max()

results_no_forfeits <- 
  str_c("match_data/results_no_forfeits_", latest_results_date, ".Rdata") %>% 
  read_rds()

team_names_by_player <- 
  bind_rows(
    results_no_forfeits %>% 
      filter(league == "SFPA", season == "Spring 2020") %>% 
      select(player = home, team = home_team),
    results_no_forfeits %>% 
      filter(league == "SFPA", season == "Spring 2020") %>% 
      select(player = away, team = away_team)
  ) %>% 
  distinct()

all_fargo_ratings <- 
  "fargo_ratings/all_fargo_ratings.Rdata" %>% 
  read_rds() %>% 
  inner_join(team_names_by_player, by = "player") %>% 
  select(-raw_rating) %>% 
  filter(date >= "2018-08-07")

initial_500_ratings <- 
  all_fargo_ratings %>% 
  group_by(player, team) %>% 
  summarize(min_date = min(date)) %>% 
  ungroup() %>% 
  filter(min_date > "2018-08-07") %>% 
  transmute(player, rating = 500, date = min_date - 7, team)

all_fargo_ratings <- 
  all_fargo_ratings %>% 
  bind_rows(initial_500_ratings)

team_list <- 
  all_fargo_ratings %>% 
  filter(date == max(date), !is.na(team)) %>% 
  group_by(team) %>% 
  summarize(rating = mean(rating)) %>% 
  arrange(desc(rating)) %>% 
  pull(team)

plot_player_ratings <- function(team_name) {
  p <- 
    all_fargo_ratings %>% 
    filter(team == team_name) %>% 
    ggplot(aes(x = date, y = rating, color = player)) +
    geom_hline(yintercept = 500, color = "white", size = 2) +
    geom_line() +
    scale_y_continuous(
      limits = c(200, 750),
      breaks = seq(200, 700, by = 100)
    ) +
    labs(
      title = str_c(team_name, " Player Ratings")
    ) +
    theme(
      axis.title = element_blank(),
      plot.title = element_text(hjust = 0.5),
      legend.title = element_blank()
    )
  
  print(p)
}
```

```{r, echo=FALSE}
walk(team_list, plot_player_ratings)
```

```{r, echo=FALSE, eval=FALSE}
# Old code for use in the early season when not everyone has played yet
# library(rvest)
# team_names_by_player <- 
#  read_html("https://www.sfpapool.org/stats/players/6") %>% 
#  html_nodes("tr") %>% 
#  html_nodes("a") %>% 
#  html_text() %>% 
#  enframe(name = NULL) %>% 
#  mutate(
#    row_type = rep(c("player", "team"), nrow(.) / 2),
#    row_number = rep(1:(nrow(.) / 2), each = 2)
#  ) %>% 
#  spread(row_type, value) %>% 
#  select(player, team) %>% 
#  mutate(
#    team = if_else(team == "Harry Harringtons v 2.0", "Harry Harringtons", team),
#    player = if_else(player == "James Horsefall", "James Horsfall", player),
#    player = if_else(player == "Humberto 'HJ' Gonzalez", "Humberto HJ Gonzalez", player),
#    player = if_else(player == "Jason XXX", "Jason Gilliland", player),
#    player = if_else(player == "Jules Tanseco", "Jukers Tanseco", player)
#  ) %>% 
#  arrange(player)
```
